refactor: fix up some checks
