+++ handlers/message_queue.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"log"
-	"math/rand"
 	"strings"
 	"time"
 
@@ -481,16 +480,4 @@ func (h *Messaging) GetPendingMessages() {
 		}
 	}
 	return
-}
-
-const charset = "abcdefghijklmnopqrstuvwxyz0123456789"
-
-var seededRand *rand.Rand = rand.New(rand.NewSource(time.Now().UnixNano()))
-
-func randString(length int) string {
-	b := make([]byte, length)
-	for i := range b {
-		b[i] = charset[seededRand.Intn(len(charset))]
-	}
-	return string(b)
 }
